perf(ssr): isolate stream cleanup runtime
